+++ forui/CHANGELOG.md
@@ -7,6 +7,9 @@
 * Add `FPopoverTagRegion`.
 * Add `FTextfield.onTap`.
 * Add `FTextfield.onTapOutside`.
+* Add `FSelectGroupController.onUpdate`.
+* Add `FValueNotifier.addValueListener(...)`.
+* Add `FValueNotifier.removeValueListener(...)`.
 
 ### Changes
 
@@ -25,10 +28,14 @@
 * **Breaking** Change `FPortalFollowerShift` to `FPortalShift`.
 * **Breaking** Remove `onChange` parameter from `FSelectTile`. This was an artifact from early prototyping that was unfortunately overlooked and did not remove.
 * **Breaking** Change `FSelectGroupController.select(...)` to `FSelectGroupController.update(...)`
+* **Breaking** Change `FSelectGroupController` to be a `ValueNotifier`.
 
 ### Fixes
 
 * Fix `FCalendar` rebuilding whenever the given `initialType` and/or `initialMonth` changes.
+
+* Resolved an issue where `FTappable` remained in a hovered or touched state when its `onPress`/`onLongPress` callbacks 
+  were nulled after being non-null.
 
 * Fix `FCalendar`'s day picker not updating when a new start and/or end date is given.